+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased][unreleased]
 
+### Changed
+
+- Move "internal use only" out of contact cards to single box at the top of the "Contacts in this trust" page
+
+## [Release-23][release-23] (production-2025-02-28.5074)
+
+### Changed
+
+- Pipeline academies pages no longer show conversion projects which are "dAO revoked"
+
 ## [Release-22][release-22] (production-2025-02-27.5055)
 
 ### Added
@@ -271,7 +281,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Updated text on "404 - Not Found" page to adhere to DfE design pattern and added the not found url to the support email template.
 
 [unreleased]:
-  https://github.com/DFE-Digital/find-information-about-academies-and-trusts/compare/production-2025-02-27.5055...HEAD
+  https://github.com/DFE-Digital/find-information-about-academies-and-trusts/compare/production-2025-02-28.5074...HEAD
 [release-1]:
   https://github.com/DFE-Digital/find-information-about-academies-and-trusts/releases/tag/production-2024-07-18.2517
 [release-2]:
@@ -315,4 +325,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 [release-21]:
   https://github.com/DFE-Digital/find-information-about-academies-and-trusts/releases/tag/production-2025-02-20.5021
 [release-22]:
-  https://github.com/DFE-Digital/find-information-about-academies-and-trusts/releases/tag/production-2025-02-27.5055
+  https://github.com/DFE-Digital/find-information-about-academies-and-trusts/releases/tag/production-2025-02-27.5055  
+[release-23]:
+  https://github.com/DFE-Digital/find-information-about-academies-and-trusts/releases/tag/production-2025-02-28.5074